requirements.txt
Flask==3.0.3
requests==2.32.3
pytest==8.3.3
Werkzeug==3.0.4
python-dotenv==1.0.0